feat(api)!: batch embeddings, mapped proxy sources, and the rest of the shape fixes
More v9-only API decisions.

Embedder had no batch method, so embedding a thousand documents meant a thousand
HTTP requests for work every backing API would have done in one — and adding the
method after tagging would break every implementor. GenerateEmbeddings is on the
interface now, implemented natively by all three providers (their responses were
already array-shaped; only the request field needed to change), with
GenerateEmbedding as the batch-of-one wrapper.

analytics.ProxySourcesConfig hardcoded IOS and Web fields, so every source an
application added was a breaking change to an exported struct here. It is a map
keyed by source name; "ios" and "web" are conventional rather than special.

outbox.RelayConfig.Dialect was documented as "must match the database.Client" —
the exact mismatch Client.Dialect() exists to make unrepresentable. The field is
gone; NewRelay and the config subpackage's NewWriter read the dialect off the
client. The SQLite-has-no-SKIP-LOCKED downgrade moved to where the dialect is
known, and narrowly, so it no longer rewrites a typo'd claim mode into a valid
one and hides it from validation.

filtering.MaxResponseSize was uint8 with a limit of 250, five below the type
ceiling, so `maxResponseSize: 300` was an unmarshal error rather than the clamp
every other over-limit value gets. It is uint16.

analytics.EventReporter.Close took no context and returned no error, so a failed
final flush of buffered events was undetectable at the one moment it matters.

llm.Capabilities.PDFs advertised a capability Part cannot express; dropped rather
than left as a flag no request can act on.

+++ dataprivacy/degraded_test.go
@@ -212,7 +212,7 @@ func TestSQLStore_CursorPagination(T *testing.T) {
 		}
 
 		filter := filtering.DefaultQueryFilter()
-		filter.MaxResponseSize = new(uint8(2))
+		filter.MaxResponseSize = new(uint16(2))
 
 		var seen []string
 
@@ -251,7 +251,7 @@ func TestSQLStore_CursorPagination(T *testing.T) {
 
 		filter := filtering.DefaultQueryFilter()
 		filter.SortBy = filtering.SortDescending
-		filter.MaxResponseSize = new(uint8(2))
+		filter.MaxResponseSize = new(uint16(2))
 
 		first, err := store.List(t.Context(), testSubject, filter)
 		must.NoError(t, err)
